// src/app/wizard-creator/enums/wizard-creator-social-media-use-cases.enum.ts
export enum WizardSocialMediaUseCases {
    InstagramCaption = 'Instagram Caption',
    LinkedinPost = 'Linkedin Post',
    FacebookPost = 'Facebook Post',
    TwitterThread = 'Twitter Tweet',
    InstagramProfileDesc = 'Instagram Profile Description',
    YoutubeDescription = 'Youtube Description',
    YoutubeVideoIdeas = 'Youtube Ideas',
    YoutubeVideoTitle = 'Youtube Title',
    HashtagGenerator = 'Hashtag Generator',
}